+++ crates/apollo-compiler/CHANGELOG.md
@@ -28,6 +28,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - **Implement `fmt::Display` for `ComponentName` - [goto-bus-stop], [pull/795]**
 - **Add `FieldDefinition::argument_by_name` and `DirectiveDefinition::argument_by_name` - [goto-bus-stop], [pull/FIXME]**
   - These methods return an argument definition by name, or `None`.
+- **Add `.lookup` methods to schema coordinates - [goto-bus-stop], [pull/FIXME]**
+  - `coord!().lookup(&schema)` returns the element at the given coordinate.
 
 [goto-bus-stop]: https://github.com/goto-bus-stop]
 [pull/795]: https://github.com/apollographql/apollo-rs/pull/795